# runtime/swagger/http.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MineAdmin API DOC",
        "description": "MineAdmin后台接口列表",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9501",
            "description": "本地服务器"
        }
    ],
    "paths": {
...
        "/system/post/index": {
            "get": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/recycle": {
            "get": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/list": {
            "get": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostListGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/save": {
            "post": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/read/{id}": {
            "get": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostRead{id}GET",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/update/{id}": {
            "put": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/delete": {
            "delete": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/realDelete": {
            "delete": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/recovery": {
            "put": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/changeStatus": {
            "put": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/numberOperation": {
            "put": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostNumberOperationPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/post/remote": {
            "post": {
                "tags": [
                    "PostController"
                ],
                "summary": "",
                "operationId": "SystemPostRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/index": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/recycle": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/save": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/read/{id}": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRead{id}GET",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/update/{id}": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/delete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/realDelete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/recovery": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/changeStatus": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/remote": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}